requirements.txt
contourpy==1.3.3
cycler==0.12.1
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
intel-openmp==2021.4.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.8.4
mkl==2021.4.0
MouseInfo==0.1.3
mpmath==1.3.0
networkx==3.6
numpy==1.26.4
opencv-python==4.9.0.80
packaging==25.0
pillow==11.3.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==2.0.1
pynput==1.8.1
pyparsing==3.2.5
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
pytweening==1.2.0
six==1.17.0
sympy==1.14.0
tbb==2021.13.1
torch==2.3.0
typing_extensions==4.15.0
